requirements.txt
pydantic~=2.9.2
python-dotenv~=1.0.1
fastapi~=0.115.2